Move NewSwapModals component to smarts
