+++ src/cow-react/pages/NewSwap/index.tsx
@@ -22,7 +22,7 @@ import { SwapButton } from 'cow-react/swap/dumb/SwapButton/SwapButton'
 import { useSwapButtonContext } from 'cow-react/swap/hooks/useSwapButtonContext'
 import { useModalIsOpen } from 'state/application/hooks'
 import { ApplicationModal } from 'state/application/reducer'
-import { NewSwapModals, NewSwapModalsProps } from 'cow-react/swap/dumb/NewSwapModals'
+import { NewSwapModals, NewSwapModalsProps } from 'cow-react/swap/smart/NewSwapModals'
 import { ConfirmSwapModalSetupProps } from 'cow-react/swap/smart/ConfirmSwapModalSetup'
 import { EthFlowProps } from 'components/swap/EthFlow'
 import AffiliateStatusCheck from 'components/AffiliateStatusCheck'
